Refactor/db sda

+++ src/project/project.controller.ts
@@ -6,7 +6,6 @@ import {
   Patch,
   Param,
   Delete,
-  Headers,
   Query,
   Req,
 } from '@nestjs/common';
@@ -17,6 +16,8 @@ import { ChatDto } from '../chat/dto/chat.dto';
 import { JwtPayloadWithRt } from '../auth/strategies/refresh.strategy';
 import { DocumentService } from '../document/document.service';
 import { AddDocumentToProjectDto } from '../document/dto/add-doc2pj.dto';
+import { InviteMemberDto } from './dto/inviteMem-project.dto';
+import { ProjectRole } from '@prisma/client';
 
 @Controller('project')
 export class ProjectController {
@@ -42,6 +43,7 @@ export class ProjectController {
     @Param('projectId') projectId: string,
     @Body() dto: AddDocumentToProjectDto,
   ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.EDITOR);
     return await this.documentService.addDocumentsToProject(
       req.user.userId,
       projectId,
@@ -56,10 +58,44 @@ export class ProjectController {
     @Param('projectId') projectId: string,
     @Body() dto: AddDocumentToProjectDto,
   ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.EDITOR);
     return await this.documentService.removeDocumentsOutProject(
       // req.user.userId,
       projectId,
       dto.documentIds,
+    );
+  }
+
+  // -- INVITE MEMBER --
+  @Post('/:projectId/members')
+  async inviteMember(
+    @Req() req: { user: JwtPayloadWithRt },
+    @Param('projectId') projectId: string,
+    @Body() dto: InviteMemberDto,
+  ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.EDITOR);
+    return await this.projectService.inviteMember(
+      req.user.userId,
+      projectId,
+      dto
+    );
+  }
+
+  // -- TOOGLE DOCS SELECTED -- 
+  @Patch(':projectId/documents/:docId/toggle')
+  async toggleDocumentSelection(
+    @Req() req: { user: JwtPayloadWithRt },
+    @Param('projectId') projectId: string,
+    @Param('docId') docId: string,
+  ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.EDITOR);
+
+    // Check exist project and doc
+
+    return await this.documentService.toggleDocumentSelection(
+      req.user.userId,
+      projectId,
+      docId,
     );
   }
 
@@ -75,6 +111,9 @@ export class ProjectController {
     @Req() req: { user: JwtPayloadWithRt },
     @Param('projectId') projectId: string,
   ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.VIEWER);
+
+
     // RETURN LIST OF CHATS IN A PROJECT
     return await this.projectService.getChatsInProject(
       req.user.userId,
@@ -88,6 +127,8 @@ export class ProjectController {
     @Req() req: { user: JwtPayloadWithRt },
     @Param('projectId') projectId: string,
   ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.VIEWER);
+
     return await this.projectService.getDocumentsInProject(
       req.user.userId,
       projectId,
@@ -120,35 +161,32 @@ export class ProjectController {
     @Query('chatId') chatId: string | undefined,
     @Body() body: ChatDto, // message
   ) {
+    await this.projectService.validateProjectAccess(req.user.userId, projectId, ProjectRole.VIEWER);
+
     body.chatId = chatId;
     body.userId = req.user.userId;
     body.projectId = projectId;
+
     // RETURN CHAT MESSAGES IN A PROJECT SPECIFIC CHAT
     return await this.projectService.chatMessageInProject(body);
   }
 
   // -- UPDATE PROJECT --
   @Patch(':id')
-  updateProject(
+  async updateProject(
+    @Req() req: { user: JwtPayloadWithRt },
     @Param('id') id: string,
     @Body() updateProjectDto: UpdateProjectDto,
   ) {
-    return this.projectService.updateProject(id, updateProjectDto);
+    await this.projectService.validateProjectAccess(req.user.userId, id, ProjectRole.EDITOR);
+    return await this.projectService.updateProject(id, updateProjectDto);
   }
 
   // -- DELETE PROJECT CASCADE --
   @Delete(':id')
-  removeProject(@Param('id') id: string) {
-    return this.projectService.removeProject(id);
+  async removeProject(@Req() req: { user: JwtPayloadWithRt }, @Param('id') id: string) {
+    await this.projectService.validateProjectAccess(req.user.userId, id, ProjectRole.OWNER);
+    return await this.projectService.removeProject(id);
   }
 
-  @Get()
-  findAll() {
-    return this.projectService.findAll();
-  }
-
-  @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.projectService.findOne(+id);
-  }
 }
